// packages/frontend/apps/web/apps/bmos-audit-web/src/components/Record/NodeList/type.ts
import { KEY } from '@/components/Layout/type';

export type nodePropsType = {
  title?: string;
  type: string;
  actived: boolean;
  key: KEY;
  used: boolean;
  icon?: string;
  rounded?: boolean;
  bordered?: boolean;
  item: ComponentNode;
  nodeKey: KEY;
  componentNumber: number | string;
};

export type NodeDataType = Pick<
  nodePropsType,
  'title' | 'actived' | 'type' | 'key'
> & { data: NodeType };

export interface OptionsType extends Record<string, any> {
  name: string;
  key: string;
  type: string;
}

export interface NodeType extends ComponentNode {
  recordItemId?: KEY;
}

export interface ComponentNode {
  componentName: string; //组件名称
  componentType: string; //组件类型: input date checkbox table
  node_type: string; //业务类型 输入 、数字、单选、领料计划、领料接收
  fieldId: KEY;
  // icon?: FunctionalComponent; //组件icon
  property?: {
    //节点参数
  };
  id?:string,
  componentDetail?: string;
  componentNumber: string | number;
}

export interface ComponentNodeList {
  basic: ComponentNode[];
  business: ComponentNode[];
}